feat(store) packages/store + packages/types adapt frontend to step-based DAG progress
