+++ apps/web/src/components/common/DynamicProgressBar.test.tsx
@@ -155,27 +155,18 @@ describe('DynamicProgressBar', () => {
           stageSlug: 'thesis',
           totalSteps: 2,
           completedSteps: 1,
+          failedSteps: 0,
           stagePercentage: 50,
           stageStatus: 'in_progress',
           stepsDetail: [
             {
               stepKey: 'plan',
               stepName: 'Plan',
-              totalJobs: 1,
-              completedJobs: 1,
-              inProgressJobs: 0,
-              failedJobs: 0,
-              stepPercentage: 100,
               status: 'completed',
             },
             {
               stepKey: 'execute',
               stepName: 'Execute',
-              totalJobs: 3,
-              completedJobs: 1,
-              inProgressJobs: 0,
-              failedJobs: 0,
-              stepPercentage: 33,
               status: 'in_progress',
             },
           ],
@@ -203,27 +194,18 @@ describe('DynamicProgressBar', () => {
           stageSlug: 'thesis',
           totalSteps: 2,
           completedSteps: 0,
+          failedSteps: 0,
           stagePercentage: 0,
           stageStatus: 'not_started',
           stepsDetail: [
             {
               stepKey: 'plan',
               stepName: 'Plan',
-              totalJobs: 0,
-              completedJobs: 0,
-              inProgressJobs: 0,
-              failedJobs: 0,
-              stepPercentage: 0,
               status: 'not_started',
             },
             {
               stepKey: 'execute',
               stepName: 'Execute',
-              totalJobs: 0,
-              completedJobs: 0,
-              inProgressJobs: 0,
-              failedJobs: 0,
-              stepPercentage: 0,
               status: 'not_started',
             },
           ],
@@ -250,17 +232,13 @@ describe('DynamicProgressBar', () => {
           stageSlug: 'thesis',
           totalSteps: 1,
           completedSteps: 1,
+          failedSteps: 0,
           stagePercentage: 100,
           stageStatus: 'completed',
           stepsDetail: [
             {
               stepKey: 'plan',
               stepName: 'Plan',
-              totalJobs: 1,
-              completedJobs: 1,
-              inProgressJobs: 0,
-              failedJobs: 0,
-              stepPercentage: 100,
               status: 'completed',
             },
           ],
@@ -269,17 +247,13 @@ describe('DynamicProgressBar', () => {
           stageSlug: 'paralysis',
           totalSteps: 1,
           completedSteps: 1,
+          failedSteps: 0,
           stagePercentage: 100,
           stageStatus: 'completed',
           stepsDetail: [
             {
               stepKey: 'plan',
               stepName: 'Plan',
-              totalJobs: 1,
-              completedJobs: 1,
-              inProgressJobs: 0,
-              failedJobs: 0,
-              stepPercentage: 100,
               status: 'completed',
             },
           ],
