+++ packages/store/src/dialecticStore.selectors.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { describe, it, expect, vi, afterEach } from 'vitest';
 import {
     selectDomains,
     selectIsLoadingDomains,
@@ -77,7 +77,6 @@ import type {
     SelectedModels,
     JobProgressEntry,
 } from '@paynless/types';
-import { STAGE_RUN_DOCUMENT_KEY_SEPARATOR } from '@paynless/types';
 
 const mockThesisStage: DialecticStage = {
     id: 's1',
@@ -96,13 +95,6 @@ const mockSynthesisStage: DialecticStage = {
     id: 's3',
     slug: 'synthesis',
     display_name: 'Synthesis',
-};
-
-const mockAntithesisStage: DialecticStage = {
-    ...mockThesisStage,
-    id: 's2',
-    slug: 'antithesis',
-    display_name: 'Antithesis',
 };
 
 describe('Dialectic Store Selectors', () => {
@@ -640,15 +632,16 @@ describe('Dialectic Store Selectors', () => {
                         stepStatuses: { doc_step: 'in_progress' },
                         documents: {},
                         jobProgress: { doc_step: jobProgressEntry },
+                        progress: { completedSteps: 0, totalSteps: 1, failedSteps: 0 },
                     },
                 },
             };
             const progress = selectUnifiedProjectProgress(stateWithEmptySelectedModels, 'session-1');
             const firstStage = progress.stageDetails[0];
             const docStep = firstStage.stepsDetail[0];
-            expect(docStep.totalJobs).toBe(3);
-            expect(docStep.completedJobs).toBe(2);
-            expect(docStep.stepPercentage).toBeCloseTo((2 / 3) * 100, 1);
+            expect(docStep.stepKey).toBe('doc_step');
+            expect(docStep.stepName).toBe('Doc Step');
+            expect(docStep.status).toBe('in_progress');
         });
     });
  
@@ -924,6 +917,7 @@ describe('Dialectic Store Selectors', () => {
               },
             },
             jobProgress: {},
+            progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
           },
         },
       };
@@ -1032,6 +1026,7 @@ describe('Dialectic Store Selectors', () => {
               },
             },
             jobProgress: {},
+            progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
           },
         },
       };
@@ -1141,6 +1136,7 @@ describe('Dialectic Store Selectors', () => {
               },
             },
             jobProgress: {},
+            progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
           },
         },
       };
@@ -1283,6 +1279,7 @@ describe('Dialectic Store Selectors', () => {
               },
             },
             jobProgress: {},
+            progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
           },
         },
       };
@@ -1839,6 +1836,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -1858,6 +1856,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -1913,6 +1912,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -1984,6 +1984,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -2060,6 +2061,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -2082,6 +2084,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -2157,6 +2160,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -2252,6 +2256,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                         'business_case': requiredDocument,
                     },
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
                 [targetProgressKey]: {
                     stepStatuses: {
@@ -2259,6 +2264,7 @@ describe('selectIsStageReadyForSessionIteration', () => {
                     },
                     documents: {},
                     jobProgress: {},
+                    progress: { completedSteps: 0, totalSteps: 0, failedSteps: 0 },
                 },
             },
         });
@@ -2266,2459 +2272,4 @@ describe('selectIsStageReadyForSessionIteration', () => {
         expect(selectIsStageReadyForSessionIteration(state, projectId, sessionId, targetStageSlug, iterationNumber)).toBe(true);
     });
 }); 
-
-// These tests assert that new selectors surface edited document resources from stageDocumentContent
-// and that legacy contribution selectors no longer drive UI renders for editable documents.
-
-describe('selectStageDocumentResource and selectEditedDocumentByKey', () => {
-    const sessionId = 'session-resource-test';
-    const stageSlug = 'thesis';
-    const iterationNumber = 1;
-    const modelId = 'model-test';
-    const documentKey = 'business_case';
-    const compositeKey = `${sessionId}:${stageSlug}:${iterationNumber}:${modelId}:${documentKey}`;
-
-    it('should return document resource content from stageDocumentContent with baselineMarkdown, timestamps, and resource id', () => {
-        const testContent: StageDocumentContentState = {
-            baselineMarkdown: 'Edited document content from resource',
-            currentDraftMarkdown: 'Edited document content from resource',
-            isDirty: false,
-            isLoading: false,
-            error: null,
-            lastBaselineVersion: {
-                resourceId: 'resource-123',
-                versionHash: 'hash-abc',
-                updatedAt: new Date('2024-01-01T00:00:00Z').toISOString(),
-            },
-            pendingDiff: null,
-            lastAppliedVersionHash: 'hash-abc',
-            sourceContributionId: null,
-            feedbackDraftMarkdown: undefined,
-            feedbackIsDirty: false,
-            resourceType: null,
-        };
-
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            stageDocumentContent: {
-                [compositeKey]: testContent,
-            },
-        };
-
-        const result = selectStageDocumentResource(state, sessionId, stageSlug, iterationNumber, modelId, documentKey);
-        
-        expect(result).toBeDefined();
-        expect(result?.baselineMarkdown).toBe('Edited document content from resource');
-        expect(result?.lastBaselineVersion?.resourceId).toBe('resource-123');
-        expect(result?.lastBaselineVersion?.updatedAt).toBe('2024-01-01T00:00:00.000Z');
-    });
-
-    it('should return edited document metadata by composite key with content and timestamps', () => {
-        const testContent: StageDocumentContentState = {
-            baselineMarkdown: 'Edited content from composite key lookup',
-            currentDraftMarkdown: 'Edited content from composite key lookup',
-            isDirty: false,
-            isLoading: false,
-            error: null,
-            lastBaselineVersion: {
-                resourceId: 'resource-456',
-                versionHash: 'hash-def',
-                updatedAt: new Date('2024-01-02T00:00:00Z').toISOString(),
-            },
-            pendingDiff: null,
-            lastAppliedVersionHash: 'hash-def',
-            sourceContributionId: null,
-            feedbackDraftMarkdown: undefined,
-            feedbackIsDirty: false,
-            resourceType: null,
-        };
-
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            stageDocumentContent: {
-                [compositeKey]: testContent,
-            },
-        };
-
-        const result = selectEditedDocumentByKey(state, compositeKey);
-        
-        expect(result).toBeDefined();
-        expect(result?.baselineMarkdown).toBe('Edited content from composite key lookup');
-        expect(result?.lastBaselineVersion?.resourceId).toBe('resource-456');
-        expect(result?.lastBaselineVersion?.updatedAt).toBe('2024-01-02T00:00:00.000Z');
-    });
-
-    it('should return undefined when document resource does not exist in stageDocumentContent', () => {
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            stageDocumentContent: {},
-        };
-
-        const result = selectStageDocumentResource(state, sessionId, stageSlug, iterationNumber, modelId, documentKey);
-        
-        expect(result).toBeUndefined();
-    });
-
-    it('should assert that resource selectors are authoritative source for editable documents, not dialectic_contributions', () => {
-        // This test asserts that for documents with entries in stageDocumentContent,
-        // the new resource selectors are the authoritative source, not legacy selectContributionById
-        
-        const testContent: StageDocumentContentState = {
-            baselineMarkdown: 'Resource-based content - authoritative source',
-            currentDraftMarkdown: 'Resource-based content - authoritative source',
-            isDirty: false,
-            isLoading: false,
-            error: null,
-            lastBaselineVersion: {
-                resourceId: 'resource-789',
-                versionHash: 'hash-ghi',
-                updatedAt: new Date('2024-01-03T00:00:00Z').toISOString(),
-            },
-            pendingDiff: null,
-            lastAppliedVersionHash: 'hash-ghi',
-            sourceContributionId: null,
-            feedbackDraftMarkdown: undefined,
-            feedbackIsDirty: false,
-            resourceType: null,
-        };
-
-        const staleContribution: DialecticContribution = {
-            id: 'contrib-stale',
-            session_id: sessionId,
-            stage: stageSlug,
-            iteration_number: iterationNumber,
-            contribution_type: documentKey,
-            is_latest_edit: true,
-            edit_version: 1,
-            created_at: new Date('2024-01-01T00:00:00Z').toISOString(),
-            updated_at: new Date('2024-01-01T00:00:00Z').toISOString(),
-            user_id: null,
-            model_id: modelId,
-            model_name: null,
-            prompt_template_id_used: null,
-            seed_prompt_url: null,
-            original_model_contribution_id: null,
-            raw_response_storage_path: null,
-            target_contribution_id: null,
-            tokens_used_input: null,
-            tokens_used_output: null,
-            processing_time_ms: null,
-            error: null,
-            citations: null,
-            file_name: null,
-            storage_bucket: null,
-            storage_path: null,
-            size_bytes: null,
-            mime_type: null,
-        };
-
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            stageDocumentContent: {
-                [compositeKey]: testContent,
-            },
-            currentProjectDetail: {
-                id: 'proj-test',
-                user_id: 'user-1',
-                project_name: 'Test Project',
-                initial_user_prompt: 'Test prompt',
-                selected_domain_id: 'dom-1',
-                dialectic_domains: { name: 'Test Domain' },
-                selected_domain_overlay_id: null,
-                repo_url: null,
-                status: 'active',
-                created_at: new Date().toISOString(),
-                updated_at: new Date().toISOString(),
-                dialectic_sessions: [{
-                    id: sessionId,
-                    project_id: 'proj-test',
-                    session_description: 'Test session',
-                    iteration_count: iterationNumber,
-                    selected_models: [{ id: modelId, displayName: modelId }],
-                    status: 'active',
-                    associated_chat_id: null,
-                    current_stage_id: 'stage-1',
-                    created_at: new Date().toISOString(),
-                    updated_at: new Date().toISOString(),
-                    user_input_reference_url: null,
-                    dialectic_contributions: [staleContribution],
-                    feedback: [],
-                }],
-                resources: [],
-                process_template_id: null,
-                dialectic_process_templates: null,
-                isLoadingProcessTemplate: false,
-                processTemplateError: null,
-                contributionGenerationStatus: 'idle',
-                generateContributionsError: null,
-                isSubmittingStageResponses: false,
-                submitStageResponsesError: null,
-                isSavingContributionEdit: false,
-                saveContributionEditError: null,
-            },
-        };
-
-        // This is the authoritative source for editable documents
-        const resourceResult = selectStageDocumentResource(state, sessionId, stageSlug, iterationNumber, modelId, documentKey);
-        expect(resourceResult?.baselineMarkdown).toBe('Resource-based content - authoritative source');
-        expect(resourceResult?.lastBaselineVersion?.resourceId).toBe('resource-789');
-        
-        // Legacy selectContributionById should NOT be used for documents that have
-        // entries in stageDocumentContent. The resource selector is authoritative.
-        // Note: selectContributionById may still work for historical lookups, but
-        // it should NOT drive UI renders for editable documents when stageDocumentContent exists.
-        const legacyResult = selectContributionById(state, staleContribution.id);
-        
-        // The test documents that for editable documents, the resource selector
-        // must be used, not the legacy contribution selector
-        // UI components should prefer resourceResult over legacyResult for editable documents
-        expect(resourceResult).toBeDefined();
-        expect(resourceResult?.baselineMarkdown).not.toBe(staleContribution.contribution_type);
-    });
-
-    it('should return document resource with dirty draft state', () => {
-        const testContent: StageDocumentContentState = {
-            baselineMarkdown: 'Original baseline content',
-            currentDraftMarkdown: 'Modified draft content',
-            isDirty: true,
-            isLoading: false,
-            error: null,
-            lastBaselineVersion: {
-                resourceId: 'resource-dirty',
-                versionHash: 'hash-dirty',
-                updatedAt: new Date('2024-01-04T00:00:00Z').toISOString(),
-            },
-            pendingDiff: null,
-            lastAppliedVersionHash: 'hash-dirty',
-            sourceContributionId: null,
-            feedbackDraftMarkdown: undefined,
-            feedbackIsDirty: false,
-            resourceType: null,
-        };
-
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            stageDocumentContent: {
-                [compositeKey]: testContent,
-            },
-        };
-
-        const result = selectStageDocumentResource(state, sessionId, stageSlug, iterationNumber, modelId, documentKey);
-        
-        expect(result).toBeDefined();
-        expect(result?.isDirty).toBe(true);
-        expect(result?.baselineMarkdown).toBe('Original baseline content');
-        expect(result?.currentDraftMarkdown).toBe('Modified draft content');
-    });
-});
-
-describe('selectValidMarkdownDocumentKeys', () => {
-    const stageSlug = 'thesis';
-    const stageWithMixedOutputs: DialecticStageRecipe = {
-        stageSlug,
-        instanceId: 'instance-mixed',
-        steps: [
-            {
-                id: 'step-markdown-1',
-                step_key: 'markdown_step_1',
-                step_slug: 'markdown-step-1',
-                step_name: 'Markdown Step 1',
-                execution_order: 1,
-                parallel_group: 1,
-                branch_key: 'branch-1',
-                job_type: 'EXECUTE',
-                prompt_type: 'Turn',
-                prompt_template_id: 'prompt-1',
-                output_type: 'assembled_document_json',
-                granularity_strategy: 'per_source_document',
-                inputs_required: [],
-                inputs_relevance: [],
-                outputs_required: [
-                    {
-                        document_key: 'draft_document_markdown',
-                        artifact_class: 'rendered_document',
-                        file_type: 'markdown',
-                    },
-                ],
-            },
-            {
-                id: 'step-json',
-                step_key: 'json_step',
-                step_slug: 'json-step',
-                step_name: 'JSON Step',
-                execution_order: 2,
-                parallel_group: 2,
-                branch_key: 'branch-2',
-                job_type: 'PLAN',
-                prompt_type: 'Planner',
-                prompt_template_id: 'prompt-2',
-                output_type: 'header_context',
-                granularity_strategy: 'all_to_one',
-                inputs_required: [],
-                inputs_relevance: [],
-                outputs_required: [
-                    {
-                        document_key: 'HeaderContext',
-                        artifact_class: 'header_context',
-                        file_type: 'json',
-                    },
-                ],
-            },
-            {
-                id: 'step-markdown-2',
-                step_key: 'markdown_step_2',
-                step_slug: 'markdown-step-2',
-                step_name: 'Markdown Step 2',
-                execution_order: 3,
-                parallel_group: 3,
-                branch_key: 'branch-3',
-                job_type: 'EXECUTE',
-                prompt_type: 'Turn',
-                prompt_template_id: 'prompt-3',
-                output_type: 'assembled_document_json',
-                granularity_strategy: 'per_source_document',
-                inputs_required: [],
-                inputs_relevance: [],
-                outputs_required: [
-                    {
-                        document_key: 'business_case_markdown',
-                        artifact_class: 'rendered_document',
-                        file_type: 'markdown',
-                    },
-                ],
-            },
-        ],
-    };
-
-    const stageWithNoMarkdownOutputs: DialecticStageRecipe = {
-        stageSlug: 'no-markdown',
-        instanceId: 'instance-no-markdown',
-        steps: [
-            {
-                id: 'step-json-only',
-                step_key: 'json_only_step',
-                step_slug: 'json-only-step',
-                step_name: 'JSON Only Step',
-                execution_order: 1,
-                parallel_group: 1,
-                branch_key: 'branch-1',
-                job_type: 'PLAN',
-                prompt_type: 'Planner',
-                prompt_template_id: 'prompt-1',
-                output_type: 'header_context',
-                granularity_strategy: 'all_to_one',
-                inputs_required: [],
-                inputs_relevance: [],
-                outputs_required: [
-                    {
-                        document_key: 'HeaderContext',
-                        artifact_class: 'header_context',
-                        file_type: 'json',
-                    },
-                ],
-            },
-        ],
-    };
-
-    it('should return a Set containing only markdown document keys, excluding non-markdown artifacts', () => {
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            recipesByStageSlug: {
-                [stageSlug]: stageWithMixedOutputs,
-            },
-        };
-
-        const result = selectValidMarkdownDocumentKeys(state, stageSlug);
-
-        expect(result).toBeInstanceOf(Set);
-        expect(result.size).toBe(2);
-        expect(result.has('draft_document_markdown')).toBe(true);
-        expect(result.has('business_case_markdown')).toBe(true);
-        expect(result.has('HeaderContext')).toBe(false);
-    });
-
-    it('should return an empty Set when no markdown outputs exist for the stage', () => {
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            recipesByStageSlug: {
-                'no-markdown': stageWithNoMarkdownOutputs,
-            },
-        };
-
-        const result = selectValidMarkdownDocumentKeys(state, 'no-markdown');
-
-        expect(result).toBeInstanceOf(Set);
-        expect(result.size).toBe(0);
-    });
-
-    it('should return an empty Set when the stage has no recipe', () => {
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            recipesByStageSlug: {},
-        };
-
-        const result = selectValidMarkdownDocumentKeys(state, 'nonexistent-stage');
-
-        expect(result).toBeInstanceOf(Set);
-        expect(result.size).toBe(0);
-    });
-
-    it('should return an empty Set when the recipe has no steps', () => {
-        const stageWithNoSteps: DialecticStageRecipe = {
-            stageSlug: 'empty-stage',
-            instanceId: 'instance-empty',
-            steps: [],
-        };
-
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            recipesByStageSlug: {
-                'empty-stage': stageWithNoSteps,
-            },
-        };
-
-        const result = selectValidMarkdownDocumentKeys(state, 'empty-stage');
-
-        expect(result).toBeInstanceOf(Set);
-        expect(result.size).toBe(0);
-    });
-
-    it('should return an empty Set when steps have no outputs_required', () => {
-        const stageWithNoOutputs: DialecticStageRecipe = {
-            stageSlug: 'no-outputs-stage',
-            instanceId: 'instance-no-outputs',
-            steps: [
-                {
-                    id: 'step-no-outputs',
-                    step_key: 'no_outputs_step',
-                    step_slug: 'no-outputs-step',
-                    step_name: 'No Outputs Step',
-                    execution_order: 1,
-                    parallel_group: 1,
-                    branch_key: 'branch-1',
-                    job_type: 'EXECUTE',
-                    prompt_type: 'Turn',
-                    prompt_template_id: 'prompt-1',
-                    output_type: 'assembled_document_json',
-                    granularity_strategy: 'per_source_document',
-                    inputs_required: [],
-                    inputs_relevance: [],
-                    outputs_required: undefined,
-                },
-            ],
-        };
-
-        const state: DialecticStateValues = {
-            ...initialDialecticStateValues,
-            recipesByStageSlug: {
-                'no-outputs-stage': stageWithNoOutputs,
-            },
-        };
-
-        const result = selectValidMarkdownDocumentKeys(state, 'no-outputs-stage');
-
-        expect(result).toBeInstanceOf(Set);
-        expect(result.size).toBe(0);
-    });
-  });
-
-  describe('selectUnifiedProjectProgress', () => {
-    const sessionId = 'session-unified';
-    const iterationNumber = 1;
-    const progressKeyForStage = (stageSlug: string): string =>
-        `${sessionId}:${stageSlug}:${iterationNumber}`;
-
-    const stage1ForUnified: DialecticStage = {
-      id: 'stage-abc',
-      slug: 'mock-stage-1',
-      display_name: 'Mock Stage 1',
-      description: 'First mock stage',
-      created_at: new Date().toISOString(),
-      default_system_prompt_id: 'sp-1',
-      expected_output_template_ids: [],
-      recipe_template_id: null,
-      active_recipe_instance_id: null,
-    };
-    const stage2ForUnified: DialecticStage = {
-      id: 'stage-def',
-      slug: 'mock-stage-2',
-      display_name: 'Mock Stage 2',
-      description: 'Second mock stage',
-      created_at: new Date().toISOString(),
-      default_system_prompt_id: 'sp-2',
-      expected_output_template_ids: [],
-      recipe_template_id: null,
-      active_recipe_instance_id: null,
-    };
-    const templateForUnified: DialecticProcessTemplate = {
-      id: 'pt-1',
-      name: 'Test Template',
-      description: 'A test template',
-      created_at: new Date().toISOString(),
-      starting_stage_id: 'stage-abc',
-      stages: [stage1ForUnified, stage2ForUnified],
-      transitions: [
-        { id: 't1', process_template_id: 'pt-1', source_stage_id: 'stage-abc', target_stage_id: 'stage-def', created_at: new Date().toISOString(), condition_description: null },
-      ],
-    };
-    const projectBaseForUnified: DialecticProject = {
-      id: 'proj-1',
-      user_id: 'user1',
-      project_name: 'Detailed Project',
-      initial_user_prompt: 'Initial Prompt Text',
-      selected_domain_id: 'domain1',
-      dialectic_domains: { name: 'Tech' },
-      selected_domain_overlay_id: 'overlay1',
-      repo_url: null,
-      status: 'active',
-      created_at: new Date().toISOString(),
-      updated_at: new Date().toISOString(),
-      dialectic_sessions: [],
-      resources: [],
-      process_template_id: 'pt1',
-      dialectic_process_templates: templateForUnified,
-      isLoadingProcessTemplate: false,
-      processTemplateError: null,
-      contributionGenerationStatus: 'idle',
-      generateContributionsError: null,
-      isSubmittingStageResponses: false,
-      submitStageResponsesError: null,
-      isSavingContributionEdit: false,
-      saveContributionEditError: null,
-    };
-
-    it('returns 0% progress for new project with no completed documents', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: {},
-        stageRunProgress: {},
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.overallPercentage).toBe(0);
-      expect(result.completedStages).toBe(0);
-      expect(result.totalStages).toBe(2);
-      expect(result.projectStatus).toBe('not_started');
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('calculates step progress as completedModels/totalModels (e.g. 1/3 = 33%)', () => {
-      const documentKey = 'doc_a';
-      const sep = STAGE_RUN_DOCUMENT_KEY_SEPARATOR;
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: documentKey, artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const baseDescriptor: StageRenderedDocumentDescriptor = {
-        descriptorType: 'rendered',
-        status: 'completed',
-        job_id: 'job-1',
-        latestRenderedResourceId: 'res-1',
-        modelId: 'model-1',
-        versionHash: 'h1',
-        lastRenderedResourceId: 'res-1',
-        lastRenderAtIso: new Date().toISOString(),
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {
-              [`${documentKey}${sep}model-1`]: { ...baseDescriptor, status: 'completed', modelId: 'model-1' },
-              [`${documentKey}${sep}model-2`]: { ...baseDescriptor, status: 'generating', modelId: 'model-2' },
-              [`${documentKey}${sep}model-3`]: { ...baseDescriptor, status: 'not_started', modelId: 'model-3' },
-            },
-            jobProgress: {
-              doc_step: { totalJobs: 3, completedJobs: 1, inProgressJobs: 2, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.totalStages).toBe(2);
-      expect(result.currentStage).not.toBeNull();
-      expect(result.projectStatus).toBe('in_progress');
-      expect(result.overallPercentage).toBeGreaterThan(0);
-      expect(result.overallPercentage).toBeLessThan(100);
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-      const firstStage = result.stageDetails[0];
-      expect(firstStage).toBeDefined();
-      expect(firstStage.stepsDetail).toBeDefined();
-      expect(firstStage.stepsDetail.length).toBeGreaterThan(0);
-      const docStep = firstStage.stepsDetail[0];
-      expect(docStep.completedJobs).toBe(1);
-      expect(docStep.totalJobs).toBe(3);
-      expect(docStep.inProgressJobs).toBe(2);
-      expect(docStep.failedJobs).toBe(0);
-      expect(docStep.stepPercentage).toBeCloseTo(100 / 3, 1);
-    });
-
-    it('one document key produced by 3 models: 1 of 3 completed yields 33% step progress', () => {
-      const documentKey = 'doc_a';
-      const sep = STAGE_RUN_DOCUMENT_KEY_SEPARATOR;
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: documentKey, artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const baseDescriptor: StageRenderedDocumentDescriptor = {
-        descriptorType: 'rendered',
-        status: 'completed',
-        job_id: 'job-1',
-        latestRenderedResourceId: 'res-1',
-        modelId: 'model-1',
-        versionHash: 'h1',
-        lastRenderedResourceId: 'res-1',
-        lastRenderAtIso: new Date().toISOString(),
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {
-              [`${documentKey}${sep}model-1`]: { ...baseDescriptor, status: 'completed', modelId: 'model-1' },
-              [`${documentKey}${sep}model-2`]: { ...baseDescriptor, status: 'generating', modelId: 'model-2' },
-              [`${documentKey}${sep}model-3`]: { ...baseDescriptor, status: 'not_started', modelId: 'model-3' },
-            },
-            jobProgress: {
-              doc_step: { totalJobs: 3, completedJobs: 1, inProgressJobs: 2, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      const firstStage = result.stageDetails[0];
-      expect(firstStage).toBeDefined();
-      expect(firstStage.stepsDetail).toBeDefined();
-      expect(firstStage.stepsDetail.length).toBeGreaterThan(0);
-      const docStep = firstStage.stepsDetail[0];
-      expect(docStep.completedJobs).toBe(1);
-      expect(docStep.totalJobs).toBe(3);
-      expect(docStep.stepPercentage).toBeCloseTo(100 / 3, 1);
-    });
-
-    it('one document key produced by 3 models: all 3 completed yields 100% step progress', () => {
-      const documentKey = 'doc_a';
-      const sep = STAGE_RUN_DOCUMENT_KEY_SEPARATOR;
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: documentKey, artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const baseDescriptor: StageRenderedDocumentDescriptor = {
-        descriptorType: 'rendered',
-        status: 'completed',
-        job_id: 'job-1',
-        latestRenderedResourceId: 'res-1',
-        modelId: 'model-1',
-        versionHash: 'h1',
-        lastRenderedResourceId: 'res-1',
-        lastRenderAtIso: new Date().toISOString(),
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {
-              [`${documentKey}${sep}model-1`]: { ...baseDescriptor, status: 'completed', modelId: 'model-1' },
-              [`${documentKey}${sep}model-2`]: { ...baseDescriptor, status: 'completed', modelId: 'model-2' },
-              [`${documentKey}${sep}model-3`]: { ...baseDescriptor, status: 'completed', modelId: 'model-3' },
-            },
-            jobProgress: {
-              doc_step: { totalJobs: 3, completedJobs: 3, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      const firstStage = result.stageDetails[0];
-      expect(firstStage).toBeDefined();
-      expect(firstStage.stepsDetail).toBeDefined();
-      expect(firstStage.stepsDetail.length).toBeGreaterThan(0);
-      const docStep = firstStage.stepsDetail[0];
-      expect(docStep.completedJobs).toBe(3);
-      expect(docStep.totalJobs).toBe(3);
-      expect(docStep.stepPercentage).toBe(100);
-    });
-
-    it('calculates stage progress as sum of step progress / total steps', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 's1',
-            step_key: 'step_a',
-            step_slug: 'step-a',
-            step_name: 'Step A',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_a', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-          {
-            id: 's2',
-            step_key: 'step_b',
-            step_slug: 'step-b',
-            step_name: 'Step B',
-            execution_order: 2,
-            parallel_group: 2,
-            branch_key: 'b2',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_b', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { step_a: 'completed', step_b: 'in_progress' },
-            documents: {
-              'doc_a': {
-                descriptorType: 'rendered',
-                status: 'completed',
-                job_id: 'j1',
-                latestRenderedResourceId: 'r1',
-                modelId: 'model-1',
-                versionHash: 'v1',
-                lastRenderedResourceId: 'r1',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-              'doc_b': {
-                descriptorType: 'rendered',
-                status: 'generating',
-                job_id: 'j2',
-                latestRenderedResourceId: 'r2',
-                modelId: 'model-1',
-                versionHash: 'v2',
-                lastRenderedResourceId: 'r2',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-            },
-            jobProgress: {
-              step_a: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 },
-              step_b: { totalJobs: 1, completedJobs: 0, inProgressJobs: 1, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.totalStages).toBe(2);
-      expect(result.currentStageSlug).toBe('mock-stage-1');
-      expect(result.overallPercentage).toBeGreaterThanOrEqual(0);
-      expect(result.overallPercentage).toBeLessThanOrEqual(100);
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('calculates overall progress as (completed stages + current stage progress) / total stages', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-def',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe1: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'i1',
-        steps: [{ id: 's1', step_key: 'step_1', step_slug: 's1', step_name: 'Step 1', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const recipe2: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-2',
-        instanceId: 'i2',
-        steps: [{ id: 's2', step_key: 'step_2', step_slug: 's2', step_name: 'Step 2', execution_order: 1, parallel_group: 1, branch_key: 'b2', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd2', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const key1 = progressKeyForStage('mock-stage-1');
-      const key2 = progressKeyForStage('mock-stage-2');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe1, 'mock-stage-2': recipe2 },
-        stageRunProgress: {
-          [key1]: {
-            stepStatuses: { step_1: 'completed' },
-            documents: {
-              'd1': {
-                descriptorType: 'rendered',
-                status: 'completed',
-                job_id: 'j1',
-                latestRenderedResourceId: 'r1',
-                modelId: 'model-1',
-                versionHash: 'v1',
-                lastRenderedResourceId: 'r1',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-            },
-            jobProgress: { step_1: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 } },
-          },
-          [key2]: {
-            stepStatuses: { step_2: 'in_progress' },
-            documents: {
-              'd2': {
-                descriptorType: 'rendered',
-                status: 'generating',
-                job_id: 'j2',
-                latestRenderedResourceId: 'r2',
-                modelId: 'model-1',
-                versionHash: 'v2',
-                lastRenderedResourceId: 'r2',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-            },
-            jobProgress: { step_2: { totalJobs: 3, completedJobs: 1, inProgressJobs: 1, failedJobs: 0 } },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.totalStages).toBe(2);
-      expect(result.completedStages).toBe(1);
-      expect(result.currentStageSlug).toBe('mock-stage-2');
-      expect(result.overallPercentage).toBeGreaterThan(50);
-      expect(result.overallPercentage).toBeLessThanOrEqual(100);
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('returns 100% when all stages complete', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-def',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe1: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'i1',
-        steps: [{ id: 's1', step_key: 'step_1', step_slug: 's1', step_name: 'Step 1', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const recipe2: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-2',
-        instanceId: 'i2',
-        steps: [{ id: 's2', step_key: 'step_2', step_slug: 's2', step_name: 'Step 2', execution_order: 1, parallel_group: 1, branch_key: 'b2', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd2', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const key1 = progressKeyForStage('mock-stage-1');
-      const key2 = progressKeyForStage('mock-stage-2');
-      const sep = STAGE_RUN_DOCUMENT_KEY_SEPARATOR;
-      const completedDoc = {
-        descriptorType: 'rendered' as const,
-        status: 'completed' as const,
-        job_id: 'j1',
-        latestRenderedResourceId: 'r1',
-        modelId: 'model-1',
-        versionHash: 'v1',
-        lastRenderedResourceId: 'r1',
-        lastRenderAtIso: new Date().toISOString(),
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe1, 'mock-stage-2': recipe2 },
-        stageRunProgress: {
-          [key1]: {
-            stepStatuses: { step_1: 'completed' },
-            documents: { [`d1${sep}model-1`]: completedDoc },
-            jobProgress: { step_1: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 } },
-          },
-          [key2]: {
-            stepStatuses: { step_2: 'completed' },
-            documents: { [`d2${sep}model-1`]: { ...completedDoc, job_id: 'j2', latestRenderedResourceId: 'r2', lastRenderedResourceId: 'r2' } },
-            jobProgress: { step_2: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 } },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.overallPercentage).toBe(100);
-      expect(result.completedStages).toBe(2);
-      expect(result.projectStatus).toBe('completed');
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('handles multi-model steps correctly (3 models = step complete only when all 3 finish)', () => {
-      const documentKey = 'doc_a';
-      const sep = STAGE_RUN_DOCUMENT_KEY_SEPARATOR;
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [{
-          id: 'step-1',
-          step_key: 'doc_step',
-          step_slug: 'doc-step',
-          step_name: 'Doc Step',
-          execution_order: 1,
-          parallel_group: 1,
-          branch_key: 'b1',
-          job_type: 'EXECUTE',
-          prompt_type: 'Turn',
-          output_type: 'assembled_document_json',
-          granularity_strategy: 'per_source_document',
-          inputs_required: [],
-          inputs_relevance: [],
-          outputs_required: [{ document_key: documentKey, artifact_class: 'rendered_document', file_type: 'markdown' }],
-        }],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const baseDoc: StageRenderedDocumentDescriptor = {
-        descriptorType: 'rendered',
-        status: 'completed',
-        job_id: 'j1',
-        latestRenderedResourceId: 'r1',
-        modelId: 'model-1',
-        versionHash: 'v1',
-        lastRenderedResourceId: 'r1',
-        lastRenderAtIso: new Date().toISOString(),
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [
-            { id: 'model-1', displayName: 'Model 1' },
-            { id: 'model-2', displayName: 'Model 2' },
-            { id: 'model-3', displayName: 'Model 3' },
-        ],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {
-              [`${documentKey}${sep}model-1`]: { ...baseDoc, status: 'completed', modelId: 'model-1' },
-              [`${documentKey}${sep}model-2`]: { ...baseDoc, status: 'generating', modelId: 'model-2' },
-              [`${documentKey}${sep}model-3`]: { ...baseDoc, status: 'not_started', modelId: 'model-3' },
-            },
-            jobProgress: {
-              doc_step: { totalJobs: 3, completedJobs: 1, inProgressJobs: 2, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.totalStages).toBe(2);
-      expect(result.projectStatus).toBe('in_progress');
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-      const firstStage = result.stageDetails[0];
-      expect(firstStage).toBeDefined();
-      expect(firstStage.stepsDetail).toBeDefined();
-      expect(firstStage.stepsDetail.length).toBeGreaterThan(0);
-      const docStep = firstStage.stepsDetail[0];
-      expect(docStep.completedJobs).toBe(1);
-      expect(docStep.totalJobs).toBe(3);
-      expect(docStep.status).toBe('in_progress');
-    });
-
-    it('handles non-model steps as 1/1 (step without model call counts as complete when step completes)', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'render-step',
-            step_key: 'render_step',
-            step_slug: 'render-step',
-            step_name: 'Render Step',
-            execution_order: 2,
-            parallel_group: 2,
-            branch_key: 'b2',
-            job_type: 'RENDER',
-            prompt_type: 'Turn',
-            output_type: 'rendered_document',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'out', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { render_step: 'completed' },
-            documents: {},
-            jobProgress: {
-              render_step: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.totalStages).toBe(2);
-      expect(result.projectStatus).toBe('in_progress');
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('mixed recipe with model and non-model steps calculates correctly', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'exec-1',
-            step_key: 'exec_step',
-            step_slug: 'exec-step',
-            step_name: 'Exec Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_x', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-          {
-            id: 'render-1',
-            step_key: 'render_step',
-            step_slug: 'render-step',
-            step_name: 'Render Step',
-            execution_order: 2,
-            parallel_group: 2,
-            branch_key: 'b2',
-            job_type: 'RENDER',
-            prompt_type: 'Turn',
-            output_type: 'rendered_document',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_x', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { exec_step: 'completed', render_step: 'completed' },
-            documents: {
-              'doc_x': {
-                descriptorType: 'rendered',
-                status: 'completed',
-                job_id: 'j1',
-                latestRenderedResourceId: 'r1',
-                modelId: 'model-1',
-                versionHash: 'v1',
-                lastRenderedResourceId: 'r1',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-            },
-            jobProgress: {
-              exec_step: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 },
-              render_step: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.totalStages).toBe(2);
-      expect(result.overallPercentage).toBeGreaterThanOrEqual(0);
-      expect(result.overallPercentage).toBeLessThanOrEqual(100);
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('returns failed status if any document has failed status', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [{
-          id: 's1',
-          step_key: 'doc_step',
-          step_slug: 'doc-step',
-          step_name: 'Doc Step',
-          execution_order: 1,
-          parallel_group: 1,
-          branch_key: 'b1',
-          job_type: 'EXECUTE',
-          prompt_type: 'Turn',
-          output_type: 'assembled_document_json',
-          granularity_strategy: 'per_source_document',
-          inputs_required: [],
-          inputs_relevance: [],
-          outputs_required: [{ document_key: 'doc_fail', artifact_class: 'rendered_document', file_type: 'markdown' }],
-        }],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'failed' },
-            documents: {
-              'doc_fail': {
-                descriptorType: 'rendered',
-                status: 'failed',
-                job_id: 'j1',
-                latestRenderedResourceId: 'r1',
-                modelId: 'model-1',
-                versionHash: 'v1',
-                lastRenderedResourceId: 'r1',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-            },
-            jobProgress: {
-              doc_step: { totalJobs: 1, completedJobs: 0, inProgressJobs: 0, failedJobs: 1 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.projectStatus).toBe('failed');
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('returns in_progress status when documents are generating', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [{
-          id: 's1',
-          step_key: 'doc_step',
-          step_slug: 'doc-step',
-          step_name: 'Doc Step',
-          execution_order: 1,
-          parallel_group: 1,
-          branch_key: 'b1',
-          job_type: 'EXECUTE',
-          prompt_type: 'Turn',
-          output_type: 'assembled_document_json',
-          granularity_strategy: 'per_source_document',
-          inputs_required: [],
-          inputs_relevance: [],
-          outputs_required: [{ document_key: 'doc_gen', artifact_class: 'rendered_document', file_type: 'markdown' }],
-        }],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {
-              'doc_gen': {
-                descriptorType: 'rendered',
-                status: 'generating',
-                job_id: 'j1',
-                latestRenderedResourceId: 'r1',
-                modelId: 'model-1',
-                versionHash: 'v1',
-                lastRenderedResourceId: 'r1',
-                lastRenderAtIso: new Date().toISOString(),
-              },
-            },
-            jobProgress: {
-              doc_step: { totalJobs: 1, completedJobs: 0, inProgressJobs: 1, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.projectStatus).toBe('in_progress');
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('returns not_started status for stages with no progress data', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: {},
-        stageRunProgress: {},
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.projectStatus).toBe('not_started');
-      expect(result.overallPercentage).toBe(0);
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-    });
-
-    it('returns stageDetails array with one entry per stage and correct StageProgressDetail shape', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_a', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'completed' },
-            documents: {},
-            jobProgress: {
-              doc_step: { totalJobs: 2, completedJobs: 2, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-
-      expect(result.stageDetails).toBeDefined();
-      expect(Array.isArray(result.stageDetails)).toBe(true);
-      expect(result.stageDetails.length).toBe(result.totalStages);
-
-      const statusValues: Array<UnifiedProjectProgress['projectStatus']> = ['not_started', 'in_progress', 'completed', 'failed'];
-      for (let i = 0; i < result.stageDetails.length; i += 1) {
-        const detail: StageProgressDetail = result.stageDetails[i];
-        expect(detail).toBeDefined();
-        expect(typeof detail.stageSlug).toBe('string');
-        expect(detail.stageSlug.length).toBeGreaterThan(0);
-        expect(typeof detail.totalSteps).toBe('number');
-        expect(typeof detail.completedSteps).toBe('number');
-        expect(detail.completedSteps).toBeGreaterThanOrEqual(0);
-        expect(detail.completedSteps).toBeLessThanOrEqual(detail.totalSteps);
-        expect(typeof detail.stagePercentage).toBe('number');
-        expect(detail.stagePercentage).toBeGreaterThanOrEqual(0);
-        expect(detail.stagePercentage).toBeLessThanOrEqual(100);
-        expect(Array.isArray(detail.stepsDetail)).toBe(true);
-        expect(statusValues).toContain(detail.stageStatus);
-        if (detail.stepsDetail.length > 0) {
-          const stepDetail = detail.stepsDetail[0];
-          expect(typeof stepDetail.totalJobs).toBe('number');
-          expect(typeof stepDetail.completedJobs).toBe('number');
-          expect(typeof stepDetail.inProgressJobs).toBe('number');
-          expect(typeof stepDetail.failedJobs).toBe('number');
-        }
-      }
-    });
-
-    it('returns correct progress when selectedModels is empty but jobs exist', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_a', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {},
-            jobProgress: {
-              doc_step: { totalJobs: 3, completedJobs: 2, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      expect(result.stageDetails).toBeDefined();
-      expect(result.stageDetails.length).toBeGreaterThan(0);
-      const firstStage = result.stageDetails[0];
-      expect(firstStage.stepsDetail.length).toBeGreaterThan(0);
-      const docStep = firstStage.stepsDetail[0];
-      expect(docStep.totalJobs).toBe(3);
-      expect(docStep.completedJobs).toBe(2);
-      expect(docStep.stepPercentage).toBeCloseTo((2 / 3) * 100, 1);
-    });
-
-    it('returns correct progress when selectedModels changes but job state unchanged', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_a', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const jobProgressEntry: JobProgressEntry = {
-        totalJobs: 2,
-        completedJobs: 2,
-        inProgressJobs: 0,
-        failedJobs: 0,
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const baseState: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'completed' },
-            documents: {},
-            jobProgress: { doc_step: jobProgressEntry },
-          },
-        },
-      };
-      const stateWithTwoModels: DialecticStateValues = { ...baseState, selectedModels: [{ id: 'model-1', displayName: 'M1' }, { id: 'model-2', displayName: 'M2' }] };
-      const stateWithOneModel: DialecticStateValues = { ...baseState, selectedModels: [{ id: 'model-1', displayName: 'M1' }] };
-      const result1: UnifiedProjectProgress = selectUnifiedProjectProgress(stateWithTwoModels, sessionId);
-      const result2: UnifiedProjectProgress = selectUnifiedProjectProgress(stateWithOneModel, sessionId);
-      const step1 = result1.stageDetails[0].stepsDetail[0];
-      const step2 = result2.stageDetails[0].stepsDetail[0];
-      expect(step1.totalJobs).toBe(step2.totalJobs);
-      expect(step1.completedJobs).toBe(step2.completedJobs);
-      expect(step1.stepPercentage).toBe(step2.stepPercentage);
-    });
-
-    it('returns stepPercentage=100 for PLAN step when jobProgress completedJobs >= 1', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'plan-step',
-            step_key: 'plan_step',
-            step_slug: 'plan-step',
-            step_name: 'Plan Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'PLAN',
-            prompt_type: 'Planner',
-            output_type: 'header_context',
-            granularity_strategy: 'all_to_one',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { plan_step: 'completed' },
-            documents: {},
-            jobProgress: {
-              plan_step: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const planStep = result.stageDetails[0].stepsDetail[0];
-      expect(planStep.stepPercentage).toBe(100);
-      expect(planStep.status).toBe('completed');
-    });
-
-    it('returns stepPercentage=0 for PLAN step when jobProgress completedJobs === 0', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'plan-step',
-            step_key: 'plan_step',
-            step_slug: 'plan-step',
-            step_name: 'Plan Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'PLAN',
-            prompt_type: 'Planner',
-            output_type: 'header_context',
-            granularity_strategy: 'all_to_one',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { plan_step: 'not_started' },
-            documents: {},
-            jobProgress: {
-              plan_step: { totalJobs: 1, completedJobs: 0, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const planStep = result.stageDetails[0].stepsDetail[0];
-      expect(planStep.stepPercentage).toBe(0);
-    });
-
-    it('returns stepPercentage=(completedJobs/totalJobs)*100 for EXECUTE step', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          {
-            id: 'step-1',
-            step_key: 'doc_step',
-            step_slug: 'doc-step',
-            step_name: 'Doc Step',
-            execution_order: 1,
-            parallel_group: 1,
-            branch_key: 'b1',
-            job_type: 'EXECUTE',
-            prompt_type: 'Turn',
-            output_type: 'assembled_document_json',
-            granularity_strategy: 'per_source_document',
-            inputs_required: [],
-            inputs_relevance: [],
-            outputs_required: [{ document_key: 'doc_a', artifact_class: 'rendered_document', file_type: 'markdown' }],
-          },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {},
-            jobProgress: {
-              doc_step: { totalJobs: 4, completedJobs: 3, inProgressJobs: 1, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const docStep = result.stageDetails[0].stepsDetail[0];
-      expect(docStep.stepPercentage).toBe((3 / 4) * 100);
-    });
-
-    it('returns stagePercentage equals average of all step percentages', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [
-          { id: 's1', step_key: 'step_a', step_slug: 'step-a', step_name: 'Step A', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] },
-          { id: 's2', step_key: 'step_b', step_slug: 'step-b', step_name: 'Step B', execution_order: 2, parallel_group: 2, branch_key: 'b2', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd2', artifact_class: 'rendered_document', file_type: 'markdown' }] },
-        ],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { step_a: 'completed', step_b: 'in_progress' },
-            documents: {},
-            jobProgress: {
-              step_a: { totalJobs: 2, completedJobs: 2, inProgressJobs: 0, failedJobs: 0 },
-              step_b: { totalJobs: 2, completedJobs: 1, inProgressJobs: 1, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const firstStage = result.stageDetails[0];
-      const expectedAvg = (100 + 50) / 2;
-      expect(firstStage.stagePercentage).toBeCloseTo(expectedAvg, 1);
-    });
-
-    it('returns overallPercentage equals (completedStages*100 + currentStagePercentage) / totalStages', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-def',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe1: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'i1',
-        steps: [{ id: 's1', step_key: 'step_1', step_slug: 's1', step_name: 'Step 1', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const recipe2: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-2',
-        instanceId: 'i2',
-        steps: [{ id: 's2', step_key: 'step_2', step_slug: 's2', step_name: 'Step 2', execution_order: 1, parallel_group: 1, branch_key: 'b2', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd2', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const key1 = progressKeyForStage('mock-stage-1');
-      const key2 = progressKeyForStage('mock-stage-2');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe1, 'mock-stage-2': recipe2 },
-        stageRunProgress: {
-          [key1]: {
-            stepStatuses: { step_1: 'completed' },
-            documents: {},
-            jobProgress: { step_1: { totalJobs: 1, completedJobs: 1, inProgressJobs: 0, failedJobs: 0 } },
-          },
-          [key2]: {
-            stepStatuses: { step_2: 'in_progress' },
-            documents: {},
-            jobProgress: { step_2: { totalJobs: 2, completedJobs: 1, inProgressJobs: 1, failedJobs: 0 } },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      expect(result.completedStages).toBe(1);
-      expect(result.totalStages).toBe(2);
-      const currentStagePct = (1 / 2) * 100;
-      const expectedOverall = (1 * 100 + currentStagePct) / 2;
-      expect(result.overallPercentage).toBeCloseTo(expectedOverall, 1);
-    });
-
-    it('returns status=failed for step when failedJobs > 0', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [{ id: 's1', step_key: 'doc_step', step_slug: 'doc-step', step_name: 'Doc Step', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'failed' },
-            documents: {},
-            jobProgress: {
-              doc_step: { totalJobs: 2, completedJobs: 1, inProgressJobs: 0, failedJobs: 1 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const docStep = result.stageDetails[0].stepsDetail[0];
-      expect(docStep.status).toBe('failed');
-    });
-
-    it('returns status=in_progress for step when inProgressJobs > 0 and failedJobs === 0', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [{ id: 's1', step_key: 'doc_step', step_slug: 'doc-step', step_name: 'Doc Step', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'in_progress' },
-            documents: {},
-            jobProgress: {
-              doc_step: { totalJobs: 2, completedJobs: 0, inProgressJobs: 2, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const docStep = result.stageDetails[0].stepsDetail[0];
-      expect(docStep.status).toBe('in_progress');
-    });
-
-    it('returns status=completed for step when completedJobs === totalJobs and totalJobs > 0', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const recipe: DialecticStageRecipe = {
-        stageSlug: 'mock-stage-1',
-        instanceId: 'inst-1',
-        steps: [{ id: 's1', step_key: 'doc_step', step_slug: 'doc-step', step_name: 'Doc Step', execution_order: 1, parallel_group: 1, branch_key: 'b1', job_type: 'EXECUTE', prompt_type: 'Turn', output_type: 'assembled_document_json', granularity_strategy: 'per_source_document', inputs_required: [], inputs_relevance: [], outputs_required: [{ document_key: 'd1', artifact_class: 'rendered_document', file_type: 'markdown' }] }],
-      };
-      const progressKey = progressKeyForStage('mock-stage-1');
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: { 'mock-stage-1': recipe },
-        stageRunProgress: {
-          [progressKey]: {
-            stepStatuses: { doc_step: 'completed' },
-            documents: {},
-            jobProgress: {
-              doc_step: { totalJobs: 3, completedJobs: 3, inProgressJobs: 0, failedJobs: 0 },
-            },
-          },
-        },
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      const docStep = result.stageDetails[0].stepsDetail[0];
-      expect(docStep.status).toBe('completed');
-    });
-
-    it('selectUnifiedProjectProgress returns correct totalStages when currentProcessTemplate has stages', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: {},
-        stageRunProgress: {},
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      expect(result.totalStages).toBe(2);
-    });
-
-    it('selectUnifiedProjectProgress returns totalStages 0 when currentProcessTemplate is null', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: null,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: {},
-        stageRunProgress: {},
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      expect(result.totalStages).toBe(0);
-    });
-
-    it('selectUnifiedProjectProgress returns correct overallPercentage based on currentProcessTemplate stages', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const projectWithSession: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithSession,
-        currentProcessTemplate: templateForUnified,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: {},
-        stageRunProgress: {},
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      expect(result.totalStages).toBe(2);
-      expect(result.overallPercentage).toBeGreaterThanOrEqual(0);
-      expect(result.overallPercentage).toBeLessThanOrEqual(100);
-    });
-
-    it('selectUnifiedProjectProgress ignores project.dialectic_process_templates and uses currentProcessTemplate', () => {
-      const session: DialecticSession = {
-        id: sessionId,
-        project_id: 'proj-1',
-        session_description: null,
-        user_input_reference_url: null,
-        iteration_count: iterationNumber,
-        selected_models: [{ id: 'model-1', displayName: 'Model 1' }],
-        status: null,
-        associated_chat_id: null,
-        current_stage_id: 'stage-abc',
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-      const projectWithTemplate: DialecticProject = {
-        ...projectBaseForUnified,
-        dialectic_sessions: [session],
-        dialectic_process_templates: templateForUnified,
-      };
-      const state: DialecticStateValues = {
-        ...initialDialecticStateValues,
-        currentProjectDetail: projectWithTemplate,
-        currentProcessTemplate: null,
-        selectedModels: [{ id: 'model-1', displayName: 'Model 1' }],
-        recipesByStageSlug: {},
-        stageRunProgress: {},
-      };
-      const result: UnifiedProjectProgress = selectUnifiedProjectProgress(state, sessionId);
-      expect(result.totalStages).toBe(0);
-    });
-  });
-
-describe('selectStageHasUnsavedChanges', () => {
-  const sessionId = 'session-unsaved-test';
-  const stageSlug = 'thesis';
-  const iterationNumber = 1;
-  const modelId = 'model-1';
-  const documentKey = 'business_case';
-  const keyPrefix = `${sessionId}:${stageSlug}:${iterationNumber}:`;
-
-  function makeContent(overrides: { isDirty?: boolean; feedbackIsDirty?: boolean }): StageDocumentContentState {
-    return {
-      baselineMarkdown: '',
-      currentDraftMarkdown: '',
-      isDirty: overrides.isDirty ?? false,
-      isLoading: false,
-      error: null,
-      lastBaselineVersion: null,
-      pendingDiff: null,
-      lastAppliedVersionHash: null,
-      sourceContributionId: null,
-      feedbackDraftMarkdown: undefined,
-      feedbackIsDirty: overrides.feedbackIsDirty ?? false,
-      resourceType: null,
-    };
-  }
-
-  it('returns { hasUnsavedEdits: false, hasUnsavedFeedback: false } when stageDocumentContent is empty', () => {
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {},
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: false, hasUnsavedFeedback: false });
-  });
-
-  it('returns { hasUnsavedEdits: true, hasUnsavedFeedback: false } when one document has isDirty: true and feedbackIsDirty: false', () => {
-    const compositeKey = `${keyPrefix}${modelId}:${documentKey}`;
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {
-        [compositeKey]: makeContent({ isDirty: true, feedbackIsDirty: false }),
-      },
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: true, hasUnsavedFeedback: false });
-  });
-
-  it('returns { hasUnsavedEdits: false, hasUnsavedFeedback: true } when one document has isDirty: false and feedbackIsDirty: true', () => {
-    const compositeKey = `${keyPrefix}${modelId}:${documentKey}`;
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {
-        [compositeKey]: makeContent({ isDirty: false, feedbackIsDirty: true }),
-      },
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: false, hasUnsavedFeedback: true });
-  });
-
-  it('returns { hasUnsavedEdits: true, hasUnsavedFeedback: true } when documents have mixed dirty states', () => {
-    const key1 = `${keyPrefix}model-a:doc1`;
-    const key2 = `${keyPrefix}model-b:doc2`;
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {
-        [key1]: makeContent({ isDirty: true, feedbackIsDirty: false }),
-        [key2]: makeContent({ isDirty: false, feedbackIsDirty: true }),
-      },
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: true, hasUnsavedFeedback: true });
-  });
-
-  it('ignores documents from different sessions (key prefix mismatch)', () => {
-    const otherSessionId = 'other-session';
-    const compositeKey = `${otherSessionId}:${stageSlug}:${iterationNumber}:${modelId}:${documentKey}`;
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {
-        [compositeKey]: makeContent({ isDirty: true, feedbackIsDirty: true }),
-      },
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: false, hasUnsavedFeedback: false });
-  });
-
-  it('ignores documents from different stages (key prefix mismatch)', () => {
-    const otherStageSlug = 'antithesis';
-    const compositeKey = `${sessionId}:${otherStageSlug}:${iterationNumber}:${modelId}:${documentKey}`;
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {
-        [compositeKey]: makeContent({ isDirty: true, feedbackIsDirty: true }),
-      },
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: false, hasUnsavedFeedback: false });
-  });
-
-  it('ignores documents from different iterations (key prefix mismatch)', () => {
-    const otherIteration = 2;
-    const compositeKey = `${sessionId}:${stageSlug}:${otherIteration}:${modelId}:${documentKey}`;
-    const state: DialecticStateValues = {
-      ...initialDialecticStateValues,
-      stageDocumentContent: {
-        [compositeKey]: makeContent({ isDirty: true, feedbackIsDirty: true }),
-      },
-    };
-    const result = selectStageHasUnsavedChanges(state, sessionId, stageSlug, iterationNumber);
-    expect(result).toEqual({ hasUnsavedEdits: false, hasUnsavedFeedback: false });
-  });
-});
